requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
argcomplete==1.10.3
Authlib==1.6.0
bcrypt==4.3.0
beautifulsoup4==4.8.2
boto3==1.38.38
botocore==1.38.38
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
compressed-rtf==1.0.7
cryptography==45.0.4
distro==1.9.0
dnspython==2.7.0
docx2txt==0.9
ebcdic==1.1.1
ecdsa==0.19.1
email_validator==2.2.0
extract-msg==0.28.7
fastapi==0.115.12
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
IMAPClient==2.1.0
itsdangerous==2.2.0
jiter==0.10.0
jmespath==1.0.1
logtail-python==0.3.3
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
msgpack==1.1.1
numpy==2.3.0
olefile==0.47
openai==1.88.0
pandas==2.3.0
passlib==1.7.4
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
pluginbase==1.0.1
psycopg==3.2.9
psycopg-binary==3.2.9
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-jose==3.5.0
python-pptx==0.6.23
pytz==2025.2
requests==2.32.4
rsa==4.9.1
s3transfer==0.13.0
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.8.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.3
xlrd==1.2.0
xlsxwriter==3.2.5
